Add '-Wsign-conversion' to sample grammar's build process
Details

 Both C & C++'s type systems reagrding signed and unsigned integers is
 a nightmare of epic proportions.  Without setting compiler options
 early in a project's lifecycle, achieving correctness in integers is
 nearly impossible.

 Consider, for example, ways to obtain the length of a file.

   o fopen(), fseek(), ftell()

     This ends up with a 'long' value; a signed value.

   o fstat(), stat()

     The length of the file is in the  'st_size' field of  'struct
     stat'.  It's of type 'off_t'.  This is defined by POSIX, and it's
     only described as a 'signed integer type'.

 Next, consider that all the bytes of the file have been loaded into
 memory, and the desire is to memcpy() all the bytes to another
 location.  The length has been determined, and it's a signed value.

 Use memcpy on it:

       void *memcpy(void *dest, const void *src, size_t n);

 It quickly becomes apparent that this won't work as desired because
 memcpy() takes an unsigned value for the number of bytes to copy, but
 the length is a signed value; the compiler will complain about the
 implicit conversion possibly changing the value of the argument.

 Consequently, to enforce sign correctness, one must necessarily cast
 from time to time.  This is not a problem, as long as the data types
 are managed consistently.

 This change updates the code that handles the generated parser to be
 correct with respect to sign & unsigned conversions.  Several
 variable's types were changed, and one (probable) Gcc defect was
 worked around in init_parser().

Testing

  source ~/lrstar/lrstar/scripts/setup \
    --build-type release \
    --bod /tmp/lrstar &&  \
  cd ${LRSTAR_DIR}

  lrstar-build -j 20 clean
  lrstar-build -j 20
  ${LRSTAR_DIR}/scripts/precommit-test

  source ~/lrstar/lrstar/scripts/setup \
    --build-type debug \
    --bod /tmp/lrstar &&  \
  cd ${LRSTAR_DIR}

  lrstar-build -j 20 clean
  lrstar-build -j 20
  ${LRSTAR_DIR}/scripts/precommit-test

   lrstar-size --compare /tmp/lrstar.unmodified/release

    Not measured.

+++ src/code/lrstar_main.cpp
@@ -21,7 +21,7 @@
 
 #if defined(LRSTAR_LINUX)
 #include <unistd.h>
-long  _filelength (int fd)
+off_t _filelength(int fd)
 {
    struct stat st;
    fstat(fd, &st);
@@ -52,7 +52,7 @@ public:
    static char   dir[256];       // Directory (folder) of file.
    static char   filename[256];  // Filename of input file.
    static char   filetype[64];   // Filetype of input file.
-   static long   filesize;
+   static size_t filesize;
    static int    filedesc;
    static int    init (char* dir, char* fn, char* ft);
    static void   term ();
@@ -86,7 +86,7 @@ char   cInput::path[256];        // Path of include file.
 char   cInput::filename[256];    // Filename of include file.
 char   cInput::filetype[64];     // Filetype of include file.
 int    cInput::filedesc;
-long   cInput::filesize;
+size_t cInput::filesize;
 
 char   cOutput::path[256];       // Path of output file.
 char   cOutput::dir[];           // Directory.
@@ -179,24 +179,28 @@ void  quit (int rc)
 int   cInput::init (char* dir, char* fn, char* ft)
 {
    long  nb;                                  // Number of bytes read.
-   strcpy (path, dir);
-   strcat (path, fn);
-   strcat (path, ft);
+
+   strcpy(path, dir);
+   strcat(path, fn);
+   strcat(path, ft);
+
    filedesc = _open (path, 0);               // Open the file.
-   if (filedesc < 0)                         // If open error.
-   {
+   if (filedesc < 0) {                       // If open error.
       printf ("File %s not found.\n", path);
       quit (1);
    }
-   filesize = _filelength(filedesc);              // Get filesize.
-   input_start = new char [filesize+6];            // Get some RAM space.
-   *input_start = '\n';                            // Put <eol> at beginning.
-   nb = _read(filedesc, input_start+1, filesize); // Read size bytes into buffer.
-   if (nb <= 0)                                    // If read error.
-   {
+
+   /* A file length cannot be negative, so the cast is safe. */
+   filesize     = static_cast<size_t>(_filelength(filedesc));
+   input_start  = new char [filesize + 6]; // Get some RAM space.
+   *input_start = '\n';                    // Put <eol> at beginning.
+
+   nb = _read(filedesc, input_start + 1, filesize); // Read size bytes into buffer.
+   if (nb <= 0) {                                 // If read error.
       printf ("Read error on file %s.\n", path);
-      quit (1);
+      quit(1);
    }
+
    input_end = input_start + nb + 1;      // Set end-of-buffer pointer.
    *input_end++ = EOL_CHAR;               // Put end-of-line here.
    *input_end++ = EOF_CHAR;               // Put first <eof> here.
